Added checks for migration rollback if on a production db. Renamed migration and seed files

+++ src/db/migrations/20180121124753_users.js
@@ -1,6 +1,7 @@
 exports.up = function(knex, Promise) {
-    return knex.schema.createTableIfNotExists('users', function (table) {
-        table.string('username', 100)
+    return knex.schema.createTableIfNotExists('users', function(table) {
+        table
+            .string('username', 100)
             .primary()
             .index('username_index');
         table.string('full_name', 200);
@@ -12,5 +13,7 @@ exports.up = function(knex, Promise) {
 };
 
 exports.down = function(knex, Promise) {
-    return knex.schema.dropTableIfExists('users');
+    if (process.env.NODE_ENV !== 'production') {
+        return knex.schema.dropTableIfExists('users');
+    }
 };
